+++ packages/camera/CHANGELOG.md
@@ -1,3 +1,22 @@
+## 0.2.0
+
+* Added support for video recording.
+* Changed the example app to add video recording.
+
+A lot of **breaking changes** in this version:
+
+Getter changes:
+ - Removed `isStarted`
+ - Renamed `initialized` to `isInitialized`
+ - Added `isRecordingVideo`
+
+Method changes:
+ - Renamed `capture` to `takePicture`
+ - Removed `start` (the preview starts automatically when `initialize` is called)
+ - Added `startVideoRecording(String filePath)`
+ - Removed `stop` (the preview stops automatically when `dispose` is called)
+ - Added `stopVideoRecording`
+
 ## 0.1.2
 
 * Fix Dart 2 runtime errors.